perf(ci): Reduce installation time in CI
Poetry caches artifacts and metadata into
its cache directory to speed things up
in subsequent installations.

Previously we were caching the virtualenv,
but Poetry was still taking a lot of time
to redownload the artifacts to build the
metadata.

Now we cache everything, so Poetry is very
fast both for locking
and installing dependencies.